fixed listBooks and minor things
