+++ cmd/api/book/service.go
@@ -1,8 +1,7 @@
 package book
 
 import (
-	"net/url"
-	"strconv"
+	"math"
 	"time"
 
 	"github.com/google/uuid"
@@ -12,16 +11,16 @@ type ServiceAPI interface {
 	ArchiveBook(id uuid.UUID) (Book, error)
 	CreateBook(bookEntry Book) (Book, error)
 	GetBook(id uuid.UUID) (Book, error)
-	ListBooks(query url.Values) (PagedBooks, error)
+	ListBooks(name string, minPrice32, maxPrice32 float32, sortBy, sortDirection string, archived bool, page, pageSize int) (PagedBooks, error)
 	UpdateBook(bookEntry Book, id uuid.UUID) (Book, error)
 }
 
 type Repository interface {
 	SetBookArchiveStatus(id uuid.UUID, archived bool) (Book, error)
-	CountRows(name string, minPrice32, maxPrice32 float32, archived bool) (int, error)
 	CreateBook(bookEntry Book) (Book, error)
 	GetBookByID(id uuid.UUID) (Book, error)
 	ListBooks(name string, minPrice32, maxPrice32 float32, sortBy, sortDirection string, archived bool, page, pageSize int) ([]Book, error)
+	ListBooksTotals(name string, minPrice32, maxPrice32 float32, archived bool) (int, error)
 	UpdateBook(bookEntry Book) (Book, error)
 }
 
@@ -57,45 +56,8 @@ type PagedBooks struct {
 	Results     []Book `json:"results"`
 }
 
-func (s *Service) ListBooks(query url.Values) (PagedBooks, error) {
-	name := query.Get("name")
-
-	var minPrice32 float32
-	minPriceStr := query.Get("min_price")
-	if minPriceStr != "" {
-		minPrice64, err := strconv.ParseFloat(minPriceStr, 32)
-		if err != nil {
-			return PagedBooks{}, ErrResponseQueryPriceInvalidFormat
-		}
-		minPrice32 = float32(minPrice64)
-	} else {
-		minPrice32 = 0
-	}
-
-	var maxPrice32 float32
-	maxPriceStr := query.Get("max_price")
-	if maxPriceStr != "" {
-		maxPrice64, err := strconv.ParseFloat(maxPriceStr, 32)
-		if err != nil {
-			return PagedBooks{}, ErrResponseQueryPriceInvalidFormat
-		}
-		maxPrice32 = float32(maxPrice64)
-	} else {
-		maxPrice32 = 9999.99 //max value to field price on db, set to: numeric(6,2)
-	}
-
-	sortBy, sortDirection, valid := extractOrderParams(query)
-	if !valid {
-		return PagedBooks{}, ErrResponseQuerySortByInvalid
-	}
-
-	archived := false
-	archivedStr := query.Get("archived")
-	if archivedStr == "true" {
-		archived = true
-	}
-
-	itemsTotal, err := s.repo.CountRows(name, minPrice32, maxPrice32, archived)
+func (s *Service) ListBooks(name string, minPrice32, maxPrice32 float32, sortBy, sortDirection string, archived bool, page, pageSize int) (PagedBooks, error) {
+	itemsTotal, err := s.repo.ListBooksTotals(name, minPrice32, maxPrice32, archived)
 	if err != nil {
 		errRepo := ErrResponse{
 			Code:    ErrResponseFromRespository.Code,
@@ -114,7 +76,7 @@ func (s *Service) ListBooks(query url.Values) (PagedBooks, error) {
 		return noBooks, nil
 	}
 
-	pagesTotal, page, pageSize, err := pagination(query, itemsTotal)
+	pagesTotal, err := pagination(page, pageSize, itemsTotal)
 	if err != nil {
 		return PagedBooks{}, err
 	}
@@ -144,4 +106,14 @@ func (s *Service) UpdateBook(bookEntry Book, id uuid.UUID) (Book, error) {
 	bookEntry.ID = id
 	bookEntry.UpdatedAt = time.Now().UTC().Round(time.Millisecond) //Atribute a new updating time to the new entry.
 	return s.repo.UpdateBook(bookEntry)
+}
+
+/*Calculates the pagination.*/
+func pagination(page, pageSize, itemsTotal int) (pagesTotal int, err error) {
+	pagesTotal = int(math.Ceil(float64(itemsTotal) / float64(pageSize)))
+	if page > pagesTotal {
+		return 0, ErrResponseQueryPageOutOfRange
+	}
+
+	return pagesTotal, nil
 }
